requirements.txt
# Core
aiohttp>=3.9,<4.0          # 키움 REST/Auth (텔레그램은 requests로 이관)
requests>=2.31              # 텔레그램 알림 (swing-trader 패턴)
websockets>=12.0,<14.0
pandas>=2.2,<3.0
numpy>=1.26,<2.0
pandas-ta>=0.3.14b
python-dotenv>=1.0
apscheduler>=3.10,<4.0
loguru>=0.7
aiosqlite>=0.19
pyyaml>=6.0

# GUI
PyQt6>=6.6
matplotlib>=3.8

# Backtest (vectorbt requires Python <3.14, install separately on 3.12)
# vectorbt>=0.26

# Dev
pytest>=8.0
pytest-asyncio>=0.23
black>=24.0
ruff>=0.3
